+++ src/main/java/com/example/lookup/repository/EstadoPedidoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EstadoPedidoRepository extends JpaRepository<EstadoPedido, Integer> {
+public interface EstadoPedidoRepository extends JpaRepository<EstadoPedido, Long> {
 }
