+++ src/main/java/com/example/lookup/repository/TiendaDistribuidoraRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TiendaDistribuidoraRepository extends JpaRepository<TiendaDistribuidora,Integer> {
+public interface TiendaDistribuidoraRepository extends JpaRepository<TiendaDistribuidora,Long> {
 }
